+++ cachebox/utils.py
@@ -33,11 +33,11 @@ def cached(
     clear_reuse: bool = False,
     info: bool = False,
 ):
-    if isinstance(cache, dict):
+    if isinstance(cache, dict) or cache is None:
         cache = _cachebox.Cache(0)
 
     if type(cache) is type or not isinstance(cache, _cachebox.BaseCacheImpl):
-        raise TypeError("we expected cachebox caches, got %r" % (type(cache).__name__,))
+        raise TypeError("we expected cachebox caches, got %r" % (cache,))
 
     def decorator(func):
         if info:
@@ -46,7 +46,7 @@ def cached(
 
             def cache_info():
                 nonlocal hits, misses
-                return _CacheInfo(hits, misses, cache.getmaxsize(), len(cache), cache.__sizeof__())
+                return _CacheInfo(hits, misses, cache.maxsize, len(cache), cache.__sizeof__())
 
             if inspect.iscoroutinefunction(func):
 
@@ -119,6 +119,118 @@ def cached(
                         pass
 
                     result = func(*args, **kwargs)
+
+                    try:
+                        return cache.setdefault(key, result)
+                    except OverflowError:
+                        return result
+
+            def cache_clear():
+                cache.clear(reuse=clear_reuse)
+
+            cache_info = None
+
+        wrapper.cache = cache
+        wrapper.cache_clear = cache_clear
+        wrapper.cache_info = cache_info
+
+        return functools.update_wrapper(wrapper, func)
+
+    return decorator
+
+
+def cachedmethod(
+    cache: _cachebox.BaseCacheImpl,
+    key_maker: typing.Callable[[tuple, dict], typing.Any] = make_key,
+    clear_reuse: bool = False,
+    info: bool = False,
+):
+    if isinstance(cache, dict) or cache is None:
+        cache = _cachebox.Cache(0)
+
+    if type(cache) is type or not isinstance(cache, _cachebox.BaseCacheImpl):
+        raise TypeError("we expected cachebox caches, got %r" % (cache,))
+
+    def decorator(func):
+        if info:
+            hits = 0
+            misses = 0
+
+            def cache_info():
+                nonlocal hits, misses
+                return _CacheInfo(hits, misses, cache.maxsize, len(cache), cache.__sizeof__())
+
+            if inspect.iscoroutinefunction(func):
+
+                async def wrapper(self, *args, **kwargs):
+                    nonlocal hits, misses
+                    key = key_maker(args, kwargs)
+                    try:
+                        result = cache[key]
+                        hits += 1
+                        return result
+                    except KeyError:
+                        misses += 1
+
+                    result = await func(self, *args, **kwargs)
+
+                    try:
+                        return cache.setdefault(key, result)
+                    except OverflowError:
+                        return result
+
+            else:
+
+                def wrapper(self, *args, **kwargs):
+                    nonlocal hits, misses
+                    key = key_maker(args, kwargs)
+                    try:
+                        result = cache[key]
+                        hits += 1
+                        return result
+                    except KeyError:
+                        misses += 1
+
+                    result = func(self, *args, **kwargs)
+
+                    try:
+                        return cache.setdefault(key, result)
+                    except OverflowError:
+                        return result
+
+            def cache_clear():
+                nonlocal hits, misses
+                cache.clear(reuse=clear_reuse)
+                hits = 0
+                misses = 0
+
+        else:
+            if inspect.iscoroutinefunction(func):
+
+                async def wrapper(self, *args, **kwargs):
+                    key = key_maker(args, kwargs)
+                    try:
+                        return cache[key]
+                    except KeyError:
+                        pass
+
+                    result = await func(self, *args, **kwargs)
+
+                    try:
+                        return cache.setdefault(key, result)
+                    except OverflowError:
+                        return result
+
+            else:
+
+                def wrapper(self, *args, **kwargs):
+                    key = key_maker(args, kwargs)
+                    try:
+                        return cache[key]
+                    except KeyError:
+                        pass
+
+                    result = func(self, *args, **kwargs)
 
                     try:
                         return cache.setdefault(key, result)