+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2024-03-14
+
+### Added
+- New decorator `cachedmethod` for class methods.
+
+### Changed
+- Now `cached` accept `None` as cache.
+
+### Fixed
+- Fix some bug
+
 ## [2.0.1] - 2024-03-09
 
 ### Added